+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 * Duplicated loading spinner on flaw lists (`OSIDB-3092`)
 * Internal comments creation fails on chrome browser (`OSIDB-3091`)
 * Auto commit edited references and ackowledgements when start editting a new one (`OSIDB-2905`)
+* Affects resolution is not updated after changing affectedness (`OSIDB-3123`)
 
 ## [2024.7.0]
 ### Changed